+++ ricaun.Revit.UI.Example/Revit/App.cs
@@ -1,6 +1,7 @@
 using Autodesk.Revit.ApplicationServices;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
+using ricaun.Revit.UI;
 using ricaun.Revit.UI.Example.Proprieties;
 using System;
 using System.Collections.Generic;
@@ -10,6 +11,80 @@ using System.Windows.Media;
 
 namespace ricaun.Revit.UI.Example.Revit
 {
+    [ApplicationLoader]
+    public class AppExample : IExternalApplication
+    {
+        private static RibbonPanel ribbonPanel;
+        public Result OnStartup(UIControlledApplication application)
+        {
+            ribbonPanel = application.CreatePanel("PanelName");
+
+            var commandButton = ribbonPanel.CreatePushButton<Commands.Command>()
+                .SetText("Command")
+                .SetToolTip("This is a tooltip.")
+                .SetLongDescription("This is a description.")
+                .SetLargeImage("/UIFrameworkRes;component/ribbon/images/revit.ico");
+
+            if (LanguageExtension.IsBrazilianPortuguese)
+            {
+                commandButton.SetText("Comando")
+                    .SetToolTip("Esta é uma dica de ferramenta.")
+                    .SetLongDescription("Esta é uma descrição.");
+            }
+
+            PulldownButton pulldownButton = ribbonPanel.CreatePulldownButton();
+            // PulldownButton pulldownButton = ribbonPanel.CreatePulldownButton(PushButtonData, PushButtonData, ...);
+            // PulldownButton pulldownButton = ribbonPanel.CreatePulldownButton("PulldownButtonName");
+            // PulldownButton pulldownButton = ribbonPanel.CreatePulldownButton("PulldownButtonName", PushButtonData, PushButtonData, ...);
+
+            SplitButton splitButton = ribbonPanel.CreateSplitButton();
+            //SplitButton splitButton = ribbonPanel.CreateSplitButton(PushButtonData, PushButtonData, ...);
+            //SplitButton splitButton = ribbonPanel.CreateSplitButton("SplitButtonName");
+            //SplitButton splitButton = ribbonPanel.CreateSplitButton("SplitButtonName", PushButtonData, PushButtonData, ...);
+
+            TextBox textBox = ribbonPanel.CreateTextBox("TextBoxName");
+            textBox.SetValue("Value");
+            textBox.SetPromptText("PromptText");
+            textBox.SetShowImageAsButton(true);
+            textBox.SetSelectTextOnFocus(true);
+
+            TextBoxData textBoxData = ribbonPanel.NewTextBoxData("TextBoxNameData");
+
+            ComboBox comboBox = ribbonPanel.CreateComboBox("ComboBoxName");
+            // ComboBox comboBox = ribbonPanel.CreateComboBox("ComboBoxName", ComboBoxMemberData, ComboBoxMemberData, ...);
+            // comboBox.AddItems(ComboBoxMemberData, ComboBoxMemberData, ...);
+
+            ComboBoxData comboBoxData = ribbonPanel.NewComboBoxData("ComboBoxDataName");
+
+            ComboBoxMemberData comboBoxMemberData = ribbonPanel.NewComboBoxMemberData("ComboBoxMemberDataName");
+            comboBoxMemberData.SetGroupName("GroupName");
+
+
+            RadioButtonGroup radioButtonGroup = ribbonPanel.CreateRadioButtonGroup("RadioButtonGroupName");
+            // RadioButtonGroup radioButtonGroup = ribbonPanel.CreateRadioButtonGroup("RadioButtonGroupName", ToggleButtonData, ToggleButtonData, ...);
+            // radioButtonGroup.AddItems(ToggleButtonData, ToggleButtonData, ...);
+
+            RadioButtonGroupData radioButtonGroupData = ribbonPanel.NewRadioButtonGroupData("RadioButtonGroupDataName");
+
+            ToggleButtonData ToggleButtonData = ribbonPanel.NewToggleButtonData("ToggleButtonDataName");
+            // ToggleButtonData ToggleButtonData = ribbonPanel.NewToggleButtonData<IExternalCommand>();
+            // ToggleButtonData ToggleButtonData = ribbonPanel.NewToggleButtonData<IExternalCommand>("ToggleButtonDataName");
+            // ToggleButtonData ToggleButtonData = ribbonPanel.NewToggleButtonData<IExternalCommand, IExternalCommandAvailability>();
+            // ToggleButtonData ToggleButtonData = ribbonPanel.NewToggleButtonData<IExternalCommand, IExternalCommandAvailability>("ToggleButtonDataName");
+
+
+
+            return Result.Succeeded;
+        }
+
+        public Result OnShutdown(UIControlledApplication application)
+        {
+            ribbonPanel?.Remove();
+            return Result.Succeeded;
+        }
+    }
+
+
     [Console]
     public class App : IExternalApplication
     {