+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.2.2] / 2022-08-08
+### Features
+- Add `IExternalCommandAvailability` on `NewToggleButtonData` Command
 ### Updated
 - Update methods to `internal`